Bruce/airtable (#9509)
* fixed linter

* cleaning

* cleaned imports

* debug crontab

* cleaning

* updated staging env

* updated other envs

---------
