+++ apps/festival/festival/src/app/dashboard/home/home.component.ts
@@ -8,8 +8,8 @@ import { OrganizationService } from '@blockframes/organization/service';
 import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-title.service';
 import { APP } from '@blockframes/utils/routes/utils';
 import { AnalyticsService } from '@blockframes/analytics/service';
-import { EventName, App, AggregatedAnalytic, createUser, isBuyer, deletedIdentifier } from '@blockframes/model';
-import { aggregate, counter, countedToAnalyticData, oneAnalyticsPerUser } from '@blockframes/analytics/utils';
+import { EventName, App, AggregatedAnalytic, createUser, isBuyer, deletedIdentifier, aggregate } from '@blockframes/model';
+import { counter, countedToAnalyticData, oneAnalyticsPerUser } from '@blockframes/analytics/utils';
 import { UserService } from '@blockframes/user/service';
 import { unique } from '@blockframes/utils/helpers';
 import { filters } from '@blockframes/ui/list/table/filters';
