+++ libs/analytics/src/lib/utils.ts
@@ -11,7 +11,6 @@ import {
   averageWatchDuration,
   EventName,
   createUser,
-  AnalyticsInteraction,
   deletedIdentifier
 } from '@blockframes/model';
 import { convertToTimeString } from '@blockframes/utils/helpers';
@@ -56,28 +55,6 @@ export function countedToAnalyticData(record: Record<string | number, number>, s
     count,
     label: scope ? toLabel(key, scope) : key
   }));
-}
-
-export function aggregate(analytics: Analytics[], data: Partial<AggregatedAnalytic> = {}) {
-  const aggregated = createAggregatedAnalytic(data);
-  for (const analytic of analytics) {
-    aggregated[analytic.name]++;
-  }
-
-  aggregated.interactions.global = aggregateInteractions(analytics);
-  aggregated.interactions.festival = aggregateInteractions(analytics.filter(a => a._meta.createdFrom === 'festival'));
-  aggregated.interactions.catalog = aggregateInteractions(analytics.filter(a => a._meta.createdFrom === 'catalog'));
-
-  return aggregated;
-}
-
-function aggregateInteractions(analytics: Analytics[]): AnalyticsInteraction {
-  const sorted = analytics.sort(sortByDate);
-  return {
-    count: sorted.length,
-    first: sorted[0] ? sorted[0]._meta.createdAt : undefined,
-    last: sorted[sorted.length - 1] ? sorted[sorted.length - 1]._meta.createdAt : undefined,
-  };
 }
 
 /**
@@ -215,10 +192,4 @@ export function toCards(aggregated: AggregatedAnalytic): MetricCard[] {
     icon: event.icon,
     selected: false
   }));
-}
-
-function sortByDate(a: Analytics, b: Analytics): number {
-  if (a._meta.createdAt.getTime() < b._meta.createdAt.getTime()) return -1;
-  if (a._meta.createdAt.getTime() > b._meta.createdAt.getTime()) return 1;
-  return 0;
 }
