+++ libs/model/src/lib/offer.ts
@@ -17,27 +17,3 @@ export interface Offer {
 export interface CrmOffer extends Offer {
   contracts: (Contract & { title: Movie; negotiation?: Negotiation })[];
 }
-
-export function offersToCrmOffers(
-  offers: Offer[],
-  contracts: Contract[],
-  titles: Movie[],
-  negotiations: (Negotiation & { contractId: string })[]
-) {
-  const crmOffers: CrmOffer[] = offers.map(o => {
-    const offerContracts = contracts.filter(c => c.offerId === o.id);
-    const crmOffer = {
-      ...o,
-      contracts: offerContracts.map(c => {
-        const contract = {
-          ...c,
-          title: titles.find(t => t.id === c.titleId),
-          negotiation: negotiations.find(n => n.contractId === c.id),
-        };
-        return contract;
-      }),
-    };
-    return crmOffer;
-  });
-  return crmOffers;
-}
